requirements.txt
# =============================================================================
# 🛡️ EmergenceLab v5
# Self-Organizing Multi-Agent Intelligence
# Entropy Constraints • Stochastic Uncertainty
# -----------------------------------------------------------------------------
# File        : requirements.txt
# Python      : 3.13+
# Environment : Research / Real-Time Simulation / Docker-Compatible
# =============================================================================


# -----------------------------------------------------------------------------
# 🌐 Web & Real-Time Communication Stack
# -----------------------------------------------------------------------------
flask==3.0.2
flask-socketio==5.3.6
eventlet==0.35.2
python-socketio==5.11.1
python-engineio==4.9.0


# -----------------------------------------------------------------------------
# 🧠 Numerical Computation & Intelligence Core
# -----------------------------------------------------------------------------
# (Aligned with Python 3.13 compatibility)
numpy>=2.1.0
pandas>=2.2.0
PyYAML==6.0.1


# -----------------------------------------------------------------------------
# 📊 Visualization, Graph Theory & Research Tooling
# -----------------------------------------------------------------------------
matplotlib>=3.9.0
networkx>=3.3


# =============================================================================
# End of Dependencies
# =============================================================================